Move group and room models into separate file
